[Section] Move create_section method to base
